services/requirements.txt
aiobotocore==2.12.0
aiohttp==3.9.1
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.13.0
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.1.0
apispec==6.3.0
appdirs==1.4.4
argcomplete==3.1.6
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
asyncssh==2.14.2
atpublic==4.0
attrs==23.1.0
autofeat==2.1.2
autopep8==2.2.0
Babel==2.13.1
backoff==2.2.1
billiard==4.2.0
blinker==1.7.0
boto3==1.34.51
botocore==1.34.51
build==1.0.3
CacheControl==0.13.1
cachelib==0.9.0
catboost==1.2.3
category-encoders==2.6.3
cattrs==23.2.3
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloudpickle==2.2.1
cmaes==0.10.0
colorama==0.4.6
colorlog==4.8.0
comm==0.2.1
configobj==5.0.8
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.2.0
crashtest==0.4.1
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.7
cycler==0.12.1
databricks-cli==0.18.0
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dictdiffer==0.9.0
dill==0.3.7
diskcache==5.6.3
distlib==0.3.7
distro==1.9.0
dnspython==2.4.2
docker==6.1.3
docutils==0.20.1
dpath==2.1.6
dulwich==0.21.6
dvc==3.48.1
dvc-data==3.14.0
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.1
dvc-s3==3.0.1
dvc-studio-client==0.20.0
dvc-task==0.3.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.0
filelock==3.13.1
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.49.0
frozenlist==1.4.0
fsspec==2024.2.0
funcy==2.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.42
google-re2==1.1
googleapis-common-protos==1.61.0
grandalf==0.8
graphviz==0.20.1
greenlet==3.0.1
grpcio==1.59.3
gto==1.7.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
hydra-core==1.3.2
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
installer==0.7.0
ipykernel==6.29.3
ipython==8.22.1
iterative-telemetry==0.0.8
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keyring==24.3.0
kiwisolver==1.4.5
kombu==5.3.5
lazy-object-proxy==1.9.0
limits==3.7.0
linkify-it-py==2.0.2
llvmlite==0.42.0
lockfile==0.12.2
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
mlflow==2.7.1
mlxtend==0.23.1
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.3.0
optuna==3.4.0
ordered-set==4.1.0
orjson==3.9.15
packaging==23.2
pandas==2.1.3
parso==0.8.3
pathspec==0.11.2
patsy==0.5.6
pendulum==2.1.2
pexpect==4.9.0
phik==0.12.4
pillow==10.2.0
Pint==0.23
pkginfo==1.9.6
platformdirs==3.11.0
plotly==5.19.0
pluggy==1.3.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prison==0.2.1
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.6
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydot==2.0.0
pyflakes==3.2.0
pygit2==1.14.1
Pygments==2.17.2
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-nvd3==0.15.0
python-slugify==8.0.1
python-telegram-bot==20.7
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
pyzmq==25.1.2
querystring-parser==1.2.4
rapidfuzz==3.5.2
referencing==0.31.1
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.0
rich-argparse==1.4.0
rpds-py==0.13.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3fs==2024.2.0
s3transfer==0.10.0
scikit-learn==1.4.1.post1
scipy==1.12.0
scmrepo==3.2.0
SecretStorage==3.3.3
semver==3.0.2
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.12
shtab==1.7.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
sqltrie==0.11.0
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.3.0
tomli==2.0.1
tomlkit==0.12.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
trove-classifiers==2023.11.29
typer==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
ujson==5.10.0
unicodecsv==0.14.1
urllib3==2.0.7
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
virtualenv==20.25.0
voluptuous==0.14.2
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.0.1
xattr==0.10.1
yarl==1.9.3
zc.lockfile==3.0.post1
zipp==3.17.0
prometheus-fastapi-instrumentator==1.1.1
prometheus-client==0.8.0